requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
bravado==10.4.1
bravado-core==5.13.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
comet-git-pure==0.19.11
comet-ml==2.0.5
configobj==5.0.6
ddt==1.2.1
everett==1.0.2
future==0.17.1
gast==0.2.2
gitdb==0.6.4
gitdb2==2.0.5
GitPython==2.1.12
google-pasta==0.1.7
grpcio==1.22.0
gym==0.13.1
h5py==2.9.0
idna==2.8
jsonpointer==2.0
jsonref==0.2
jsonschema==3.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-rl==0.4.2
Markdown==3.1.1
pypi
mock==3.0.5
monotonic==1.5
msgpack-python==0.5.6
neptune-client==0.4.3.
netifaces==0.10.9
nolds
numpy==1.16.4
nvidia-ml-py3==7.352.0
oauthlib==3.0.2
pandas==0.25.0
Pillow==6.1.0
protobuf==3.9.0
psutil==5.6.3
pyglet==1.3.2
PyJWT==1.7.1
pyreadline==2.1
pyrsistent==0.15.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
rfc3987==1.3.8
scikit-learn==0.21.3
scipy==1.3.0
simplejson==3.16.0
six==1.12.0
smmap==0.9.0
smmap2==2.0.5
strict-rfc3339==0.7
swagger-spec-validator==2.4.3
tensorboard==1.14.0
tensorflow==1.13.1
tensorflow-estimator==1.14.0
termcolor==1.1.0
typing-extensions==3.7.4
urllib3==1.25.3
webcolors==1.9.1
websocket-client==0.56.0
Werkzeug==0.15.5
wincertstore==0.2
wrapt==1.11.2
wurlitzer==1.0.3
